# docs/swagger.yaml
basePath: /
definitions:
  parser.Alerts:
    properties:
      ends:
        type: string
      enemyWaves:
        type: integer
      id:
        type: string
      maxEnemyLevel:
        type: integer
      minEnemyLevel:
        type: integer
      missionFaction:
        type: string
      missionLocation:
        type: string
      missionType:
        type: string
      rewardCredits:
        type: integer
      rewardItem:
        type: string
      rewardItemMany:
        type: string
      rewardItemManyCount:
        type: integer
      started:
        type: string
    type: object
  parser.DarvoDeals:
    properties:
      dealPrice:
        type: integer
      discountPercent:
        type: integer
      ends:
        type: string
      id:
        type: string
      item:
        type: string
      itemtest:
        type: string
      price:
        type: integer
      sold:
        type: integer
      start:
        type: string
      stock:
        type: integer
    type: object
  parser.News:
    properties:
      date:
        type: string
      id:
        type: string
      image:
        type: string
      message:
        type: string
      url:
        type: string
    type: object
host: api.tenno.dev
info:
  contact: {}
  description: Tenno.dev  APIs
  title: Tenno.dev  APIs
  version: "0.1"
paths:
  /{platform}/alerts:
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Platform
        in: path
        name: platform
        required: true
        type: string
      - description: lang select
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/parser.Alerts'
      summary: Show current Alerts
      tags:
      - Show Alertsdata
  /{platform}/darvo:
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Platform
        in: path
        name: platform
        required: true
        type: string
      - description: lang select
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/parser.DarvoDeals'
      summary: Show active  Darvo Deals
      tags:
      - Show DarvoDeals
  /{platform}/news:
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Platform
        in: path
        name: platform
        required: true
        type: string
      - description: lang select
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/parser.News'
      summary: Show current News
      tags:
      - Show Newsdata
swagger: "2.0"